+++ shared/src/commonMain/kotlin/com/dishdiscoverers/foodrecipe/dongguo/FetchAPI.kt
@@ -56,14 +56,3 @@ class FetchRecipe {
     }
 }
 
-@Serializable
-data class RocketLaunch(
-    @SerialName("flight_number")
-    val flightNumber: Int,
-    @SerialName("name")
-    val missionName: String,
-    @SerialName("date_utc")
-    val launchDateUTC: String,
-    @SerialName("success")
-    val launchSuccess: Boolean?,
-)
